pyproject.toml
[project]
name = "runlet-py"
version = "0.1.0"
description = ""
authors = [
    {name = "TheAppleKingy",email = "pridanov.ali.job@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "fastapi (>=0.125.0,<0.126.0)",
    "sqlalchemy (>=2.0.45,<3.0.0)",
    "alembic (>=1.17.2,<2.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "pytest (>=9.0.2,<10.0.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "aioredis (>=2.0.1,<3.0.0)",
    "freezegun (>=1.5.5,<2.0.0)",
    "uvicorn (>=0.38.0,<0.39.0)",
    "pytest-mock (>=3.15.1,<4.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "aio-pika (>=9.5.8,<10.0.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "sqlalchemy-utils (>=0.42.1,<0.43.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
